Check if last shown Emoji category ID is still valid
When upgrading to a new version of APK we may find that the stored Emoji
category ID is no longer applicable. For example it happened when we
brought in the new Unicode 8.0 Emojis.

Also restore the "Symbols" emoji category on pre-kitkat devices which
was accidentally dropped when bringing in Unicode 8.0 Emojis.

b/25972978

Change-Id: I91c044603b0aac8757cb8597d3af995f84b822f3