Experimentally specify languageTag attribute.
With previous CLs [1][2], now we can associate a BCP 47 language tag for
each InputMethodSubtype in XML resource file by "languageTag" attribute.
In order to test that the functionality, we start using "languageTag"
for some subtypes.

Note that specifying "languageTag" for all the existing subtypes is
beyond the goal of this CL, which should be handled in subsequent CLs.

Here is the list of subtypes that start having "languageTag" attribute.
  - android:imeSubtypeLocale="en_US" -> android:languageTag="en-US"
  - android:imeSubtypeLocale="en_GB" -> android:languageTag="en-GB"
  - android:imeSubtypeLocale="fr   " -> android:languageTag="fr"
  - android:imeSubtypeLocale="fr_CA" -> android:languageTag="fr-CA"
  - android:imeSubtypeLocale="fr_CH" -> android:languageTag="fr-CH"
  - android:imeSubtypeLocale="tl"    -> android:languageTag="fil"

  [1]: I77db5b99a7cf745d800db75baf135bb60ad04820
       8d6eeb01df72891acd3aa75e64aa1595a41cc96e
  [2]: I251d3d999afd13c0d618f2cb59e8ed3d47f21c98
       b8456a6a483ce62c81b92f613561fb761be0f3e8

Bug: 22858221
Change-Id: I37cb9ce196f2e23589e60ce34475504405778bbb