Use android:languageTag for Serbian-Latin layout.
Serbian-Latin is one of subtypes that our APIs could not easily deal
with their locale.

Now we can safely rely on BCP 47 language tag hence we should do.  See
my previous CLs [1][2][3] for details.

Note that this CL is supposed to have no effect on the behavior of
LatinIME because we internall remap "sr-Latn" to "sr_ZZ" [4].

 [1]: I77db5b99a7cf745d800db75baf135bb60ad04820
      8d6eeb01df72891acd3aa75e64aa1595a41cc96e
 [2]: I251d3d999afd13c0d618f2cb59e8ed3d47f21c98
      b8456a6a483ce62c81b92f613561fb761be0f3e8
 [3]: I37cb9ce196f2e23589e60ce34475504405778bbb
      f6997344e6e31ee6e7b95aec6853e0ded00d4083
 [4]: I93ff0c75b3687bb1b913f451b9eb5f2820beefbc
      31a3f07c2117180df524d6b2e78089897c8ea807

Bug: 27348943
Change-Id: Icd0667c40c6f1310699f3deb1234b7861ec2fa24