Re-introduce <uses-sdk> to the main AndroidManifest.xml
This logically reverts my previous CL [1], which moved <uses-sdk />
from AndroidManifest.xml to AndroidManifest_SdkVersion.xml, which was
later accidentally removed during Android.bp migration [2].

Since gradle support in LatinIME is still experimental and not used in
official build, it is OK to keep it broken for a while.  A more
important problem right now is letting LatinIME target API 28.

As the first step to let LatinIME target SDK 28, <uses-sdk> will be
revived in AndroidManifest.xml

 [1]: Ib673bca5a31b2f95329c9310a127ec0701bd8fdc
      9cd42f39eb31d4f0b9cb3a795025e8ab3663f3c8
 [2]: Ib8867d3b74f09fc1d9f95adc9a49a81ac0f7f054
      22ebec6e2e4bdf3f18023907e649c68ab8b3c7a3

Bug: 110741422
Bug: 119489995
Test: manually verified that LatinIME still is functional
Change-Id: I8b04caa56fe2431c07e076fb6f97c560d5da8838