Move MODULE_LICENSE_APACHE2 to the project top dir
In general files in LatinIME project should be Apache 2 license
(unless some exceptional note is there).  This is not limited to Java
source files.

Hence moving MODULE_LICENSE_APACHE2 [1] to the project top directory
would make much more sense to avoid possible confusions.

NOTICE file is a bit complicated because it also needs to be copied
into the system image.  Since the actual build rule for LatinIME.apk
is placed as java/Android.bp, and
  notice: "../NOTICE"
is something that is not supported / recommended, this CL just copies
NOTICE file from java/NOTICE to the top level directory.

Note also that NOTICE for LatinIME has a special notation [2].

 [1]: Originally added in the following CL for Bug 2548782
      If50b2e9e9cc4b4876b0d047fc9f34e2d537a6da4
      443c360d0afdbab091994244f045f4756feaf2b4
 [2]: Originally added in the following CL for Bug 7285172
      Ib4f4e433850d3f58e0a7cef205be4dc59c7f28c4
      49d301e9ad470a83048e0a5093cd88e54dd491be

Test: Locally verified as follows:
  lunch aosp_taimen-userdebug
  make -j
  cat $OUT/obj/NOTICE_FILES/src/system/app/LatinIME/LatinIME.apk.txt
Change-Id: I0ca84c46125de7f86f9e26906e107ac9ad7af177