Using IME context to inflate layout from S_V2
With CL[1],[2] to migrate InputMethodService as the subclass of the new
introduced class WindowProviderService in S_V2, IME context resources
can be managed by associating the window container of IME window when
its display/window configuration changed.

So we can get rid of createDisplayContext logic from S_V2 with gated
by SDK version and refining the method of get IME context with
documentation to make it clear.

[1]: Ie565e30ed5dd3f2cfe27355a6dded76dc3adc14b
[2]: I64a1614f32d097785915f6105b1813a929e0fe32

Bug: 213118079
Bug: 133825283
Test: manual with below steps
 1) adb install -r EditTextVariations.apk
 2) tapas LatinIME
 3) make
 4) adb install -r out/target/product/generic/system/app/LatinIME/\
        LatinIME.apk
 5) adb shell ime enable com.android.inputmethod.latin/.LatinIME
 6) adb shell ime set com.android.inputmethod.latin/.LatinIME
 5) Enable screen auto-rotation
 7) Launch EditTextVariations from launcher's shortcut
 8) Tap the first EditText field to show IME
 9) Rotate the device to the landscape mode
 10) Expect the IME should not be shrunk

Change-Id: If2cc1c5bdb257a9c0af653fa7157cf781a90bf1d