Update keyboard texts for many languages to separate misc keys
This was done by looking at characters listed on the Wikipedia language recognition chart[0] for each language, finding the last character in the array that the language uses and splitting the rest off into misc. This was partially automated, but there is potential that there may be errors for some languages. All of the keys should still be there if misc keys are enabled, so the keyboard should still be usable for all affected languages even if there were errors, if misc keys are enabled.

[0] https://en.wikipedia.org/wiki/Wikipedia:Language_recognition_chart